+++ docs/html/search/all_7.js
@@ -1,11 +1,12 @@
 var searchData=
 [
-  ['incrementmod_32',['incrementMod',['../arithmetic_8h.html#a0eefb1f15b20e28182334fb15ab361e0',1,'ace_common']]],
-  ['incrementmodoffset_33',['incrementModOffset',['../arithmetic_8h.html#a7340464d30e60b50d608cbc22ca4bea7',1,'ace_common']]],
-  ['isnull_34',['isNull',['../classace__common_1_1FCString.html#a08cbc30fffe6c62228c3985465a15db3',1,'ace_common::FCString']]],
-  ['isreversesorted_35',['isReverseSorted',['../isSorted_8h.html#ac102351f17141f6de4b924e0c965eb2b',1,'ace_common']]],
-  ['isreversesortedbykey_36',['isReverseSortedByKey',['../isSorted_8h.html#a085872c61c275d0a4a9db0a450292ca4',1,'ace_common']]],
-  ['issorted_37',['isSorted',['../isSorted_8h.html#a8563eb5d0660b39f07cda5a03ee81363',1,'ace_common']]],
-  ['issorted_2eh_38',['isSorted.h',['../isSorted_8h.html',1,'']]],
-  ['issortedbykey_39',['isSortedByKey',['../isSorted_8h.html#af33b9732411d0fe1340f27cc9ba07507',1,'ace_common']]]
+  ['incrementmod_34',['incrementMod',['../arithmetic_8h.html#a0eefb1f15b20e28182334fb15ab361e0',1,'ace_common']]],
+  ['incrementmodoffset_35',['incrementModOffset',['../arithmetic_8h.html#a7340464d30e60b50d608cbc22ca4bea7',1,'ace_common']]],
+  ['ishexchar_36',['isHexChar',['../backslash__x__encoding_8h.html#a9d4a1532b7b39c6dae0c466df9320df8',1,'ace_common']]],
+  ['isnull_37',['isNull',['../classace__common_1_1FCString.html#a08cbc30fffe6c62228c3985465a15db3',1,'ace_common::FCString']]],
+  ['isreversesorted_38',['isReverseSorted',['../isSorted_8h.html#ac102351f17141f6de4b924e0c965eb2b',1,'ace_common']]],
+  ['isreversesortedbykey_39',['isReverseSortedByKey',['../isSorted_8h.html#a085872c61c275d0a4a9db0a450292ca4',1,'ace_common']]],
+  ['issorted_40',['isSorted',['../isSorted_8h.html#a8563eb5d0660b39f07cda5a03ee81363',1,'ace_common']]],
+  ['issorted_2eh_41',['isSorted.h',['../isSorted_8h.html',1,'']]],
+  ['issortedbykey_42',['isSortedByKey',['../isSorted_8h.html#af33b9732411d0fe1340f27cc9ba07507',1,'ace_common']]]
 ];