+++ docs/html/search/files_1.js
@@ -1,4 +1,5 @@
 var searchData=
 [
-  ['binarysearch_2eh_101',['binarySearch.h',['../binarySearch_8h.html',1,'']]]
+  ['backslash_5fx_5fencoding_2eh_104',['backslash_x_encoding.h',['../backslash__x__encoding_8h.html',1,'']]],
+  ['binarysearch_2eh_105',['binarySearch.h',['../binarySearch_8h.html',1,'']]]
 ];