+++ src/AceCommon.h
@@ -51,27 +51,35 @@ SOFTWARE.
 #endif
 
 #include "arithmetic/arithmetic.h"
+
 #include "pstrings/pstrings.h"
+#include "fstrings/FCString.h"
+#include "fstrings/FlashString.h"
+#include "kstrings/KString.h"
+#include "cstrings/copyReplace.h"
+
 #include "print_str/PrintStr.h"
+
 #include "print_utils/printPadTo.h"
 #include "print_utils/printfTo.h"
+#include "print_utils/printReplaceTo.h"
+#include "print_utils/printIntAsFloat.h"
+
 #include "timing_stats/TimingStats.h"
 #include "timing_stats/GenericStats.h"
+
 #include "url_encoding/url_encoding.h"
-#include "fstrings/FCString.h"
+#include "backslash_x_encoding/backslash_x_encoding.h"
+
 #include "hash/djb2.h"
-#include "kstrings/KString.h"
+
 #include "algorithms/binarySearch.h"
 #include "algorithms/isSorted.h"
 #include "algorithms/linearSearch.h"
-#include "cstrings/copyReplace.h"
-#include "print_utils/printReplaceTo.h"
-#include "fstrings/FlashString.h"
-#include "print_utils/printIntAsFloat.h"
 #include "algorithms/reverse.h"
 
 // Version format: "xx.yy.zz" => xxyyzz (without leading 0)
-#define ACE_COMMON_VERSION 10407
-#define ACE_COMMON_VERSION_STRING "1.4.7"
+#define ACE_COMMON_VERSION 10500
+#define ACE_COMMON_VERSION_STRING "1.5.0"
 
 #endif